update to fix a deploy bug

+++ src/BusinessLayer/Service/UserService.cs
@@ -83,7 +83,7 @@ namespace AlwaysMoveForward.AnotherBlog.BusinessLayer.Service
 
             return user;
         }
-        public void Delete(int userId)
+        public void Delete(long userId)
         {
             AnotherBlogUser targetUser = this.UserRepository.GetById(userId);
 
@@ -102,7 +102,7 @@ namespace AlwaysMoveForward.AnotherBlog.BusinessLayer.Service
             return this.UserRepository.GetAll();
         }
 
-        public AnotherBlogUser GetById(int userId)
+        public AnotherBlogUser GetById(long userId)
         {
             return this.UserRepository.GetById(userId);
         }      
@@ -112,7 +112,7 @@ namespace AlwaysMoveForward.AnotherBlog.BusinessLayer.Service
             return this.UserRepository.GetBlogWriters(targetBlog.Id);
         }
 
-        public AnotherBlogUser AddBlogRole(int userId, int blogId, RoleType.Id roleId)
+        public AnotherBlogUser AddBlogRole(long userId, long blogId, RoleType.Id roleId)
         {
             AnotherBlogUser retVal = null;
 
@@ -140,7 +140,7 @@ namespace AlwaysMoveForward.AnotherBlog.BusinessLayer.Service
             return retVal;
         }
 
-        public AnotherBlogUser RemoveBlogRole(int userId, int blogId)
+        public AnotherBlogUser RemoveBlogRole(long userId, long blogId)
         {
             AnotherBlogUser retVal = null;
